+++ pandasdmx/util.py
@@ -181,6 +181,45 @@ class MaybeCachedSession(type):
         return super().__new__(cls, name, (base,), dct)
 
 
+KT = TypeVar("KT")
+VT = TypeVar("VT")
+
+log = logging.getLogger(__name__)
+
+__all__ = [
+    "BaseModel",
+    "DictLike",
+    "compare",
+    "dictlike_field",
+    "only",
+    "summarize_dictlike",
+    "validate_dictlike",
+    "validator",
+]
+
+
+class BaseModel(pydantic.BaseModel):
+    """Common settings for :class:`pydantic.BaseModel` in :mod:`sdmx`."""
+
+    class Config:
+        copy_on_model_validation = False
+        validate_assignment = True
+
+
+class MaybeCachedSession(type):
+    """Metaclass to inherit from :class:`requests_cache.CachedSession`, if available.
+
+    If :mod:`requests_cache` is not installed, returns :class:`requests.Session` as a
+    base class.
+    """
+
+    def __new__(cls, name, bases, dct):
+        base = (
+            requests.Session if not HAS_REQUESTS_CACHE else requests_cache.CachedSession
+        )
+        return super().__new__(cls, name, (base,), dct)
+
+
 class DictLike(dict, typing.MutableMapping[KT, VT]):
     """Container with features of a dict & list, plus attribute access."""
 
@@ -254,7 +293,7 @@ class DictLike(dict, typing.MutableMapping[KT, VT]):
             if error:
                 raise ValidationError([error], self.__class__)
             else:
-                return v.popitem()
+                return (key, value)
 
     def compare(self, other, strict=True):
         """Return :obj:`True` if `self` is the same as `other`.